refactor(uat-assessment): modularize assessment-template.js into 8 section modules
SD-LEO-REFAC-UAT-ASSESS-003: Refactor UAT Assessment Template

- Split 1252 LOC single template string into 8 maintainable section modules:
  - sections/header.js (16 LOC) - Metadata and test info
  - sections/page-intent.js (96 LOC) - Purpose and user flow
  - sections/backend-evaluation.js (144 LOC) - Services and data layer
  - sections/ui-ux-assessment.js (149 LOC) - Design system and accessibility
  - sections/integration-check.js (51 LOC) - Frontend-backend mapping
  - sections/sub-agent-responsibilities.js (230 LOC) - Task assignments
  - sections/testing-scope.js (535 LOC) - Manual UAT checklists
  - sections/summary.js (80 LOC) - Priority actions and QA notes
  - sections/index.js (57 LOC) - Section assembler

- Convert original file to 35 LOC backward-compatible re-export wrapper
- Add comprehensive test suite (14 tests) verifying:
  - Backward compatibility for existing imports
  - All section exports
  - Assembled template integrity

All consumers continue to work without changes.
