+++ scripts/modules/uat-assessment/assessment-template.js
@@ -1,1252 +1,35 @@
 /**
  * Chairman Console Comprehensive UI/UX Quality Assessment Template
- * Extracted from update-uat-test-nav-001-assessment.js
+ *
+ * REFACTORED: This file now re-exports from modular sections.
+ * The template has been split into 8 maintainable section files:
+ *
+ * - sections/header.js - Metadata and test info
+ * - sections/page-intent.js - Purpose and user flow (Section 1)
+ * - sections/backend-evaluation.js - Services and data layer (Section 2)
+ * - sections/ui-ux-assessment.js - Design system and accessibility (Section 3)
+ * - sections/integration-check.js - Frontend-backend mapping (Section 4)
+ * - sections/sub-agent-responsibilities.js - Task assignments (Section 5)
+ * - sections/testing-scope.js - Manual UAT checklists
+ * - sections/summary.js - Priority actions and QA notes
+ *
+ * Original file: 1252 LOC (single template string)
+ * Refactored: 8 section modules + assembler (~150-250 LOC each)
  *
  * @module assessment-template
  */
 
-export const comprehensiveAssessment = `# Chairman Console - Comprehensive UI/UX Quality Assessment (Systematic UAT)
-
-**Test ID**: TEST-NAV-001
-**Section**: chairman-console
-**Route**: \`/chairman\`
-**Assessment Date**: 2025-10-01
-**Assessment Type**: Full Page Systematic UAT Evaluation
-
----
-
-## 1. PAGE INTENT & CONTEXT
-
-### Primary Purpose
-The **Chairman Console** serves as the strategic command center and default entry point for the EHG application, providing executive-level oversight of the entire venture portfolio with real-time intelligence and cross-company analytics.
-
-### Strategic Position in Application
-- **Default Landing Page**: All authenticated users redirect to \`/chairman\` after login
-- **Hub Pattern**: Acts as gateway to all other application areas (Ventures, Portfolios, EVA, Analytics)
-- **First Item**: Top position in "Main Navigation" sidebar category
-- **Executive Focus**: Designed for C-level decision-making with high-level aggregated views
-
-### User Flow Architecture
-\`\`\`
-Authentication Flow:
-  Login -> ProtectedRoute -> AuthenticatedLayout -> /chairman
-
-Page Structure:
-  Chairman Console (/chairman)
-  +- Header
-  |  +- SidebarTrigger (mobile menu)
-  |  +- GlobalSearch (HeaderSearch)
-  |  +- BreadcrumbNavigation
-  |  +- Header Actions
-  |     +- KeyboardShortcuts
-  |     +- CompanySelector
-  |     +- DarkModeToggle
-  |     +- UserMenu
-  |
-  +- Executive Overview (Always Visible)
-  |  +- Portfolio Value
-  |  +- Active Ventures
-  |  +- Success Rate
-  |  +- At Risk
-  |
-  +- Tabbed Dashboard (6 Views)
-     +- Overview (Default Tab)
-     |  +- Performance Drive Cycle
-     |  +- AI Strategic Insights
-     |  +- Priority Alerts
-     |  +- Synergy Opportunities
-     |  +- Portfolio Performance Summary
-     +- Portfolio -> VenturePortfolioOverview
-     +- KPIs -> StrategicKPIMonitor
-     +- Financial -> FinancialAnalytics
-     +- Operations -> OperationalIntelligence
-     +- Intelligence -> AIInsightsEngine + Decision Support
-\`\`\`
-
-### Relationship to Other Pages
-
-| Target Page | Relationship | Navigation Path |
-|-------------|--------------|-----------------|
-| EVA Assistant | Peer feature | Sidebar -> EVA Assistant |
-| Ventures | Detail view | Overview cards -> Ventures count -> /ventures |
-| Portfolios | Detail view | Sidebar -> Portfolios |
-| Analytics | Deep dive | Via embedded charts -> Full analytics |
-| Reports | Export destination | "Export Report" button |
-| Settings | Configuration | UserMenu -> Settings |
-| Notifications | Alert details | "View All Alerts" button |
-
-### Component Dependencies (17 components)
-
-#### Layout Components
-1. **AuthenticatedLayout** - Overall page shell with sidebar + header
-2. **ModernNavigationSidebar** - Left navigation with chairman-console as first item
-3. **BreadcrumbNavigation** - Current location indicator
-4. **FloatingEVAAssistant** - Persistent AI assistant overlay
-
-#### Header Components
-5. **SidebarTrigger** - Mobile menu toggle
-6. **HeaderSearch** - Global search functionality
-7. **CompanySelector** - Multi-company filtering
-8. **KeyboardShortcuts** - Shortcut reference dialog
-9. **DarkModeToggle** - Theme switcher
-10. **UserMenu** - Profile + logout dropdown
-
-#### Dashboard Components
-11. **ExecutiveOverviewCards** - 4 metric cards (Portfolio Value, Active Ventures, Success Rate, At Risk)
-12. **PerformanceDriveCycle** - Performance visualization widget
-13. **AIInsightsEngine** - AI-powered strategic insights
-14. **ExecutiveAlerts** - Priority alerts with real-time updates
-15. **SynergyOpportunities** - Cross-venture synergy detection
-16. **VenturePortfolioOverview** - Detailed portfolio analysis (Portfolio tab)
-17. **StrategicKPIMonitor** - KPI tracking dashboard (KPIs tab)
-18. **FinancialAnalytics** - Financial metrics and charts (Financial tab)
-19. **OperationalIntelligence** - Operational metrics (Operations tab)
-
----
-
-## 2. BACKEND EVALUATION
-
-### FULLY CONNECTED Services
-
-#### Authentication & Authorization
-- **Component**: \`ProtectedRoute\` wrapper
-- **Service**: Supabase Auth
-- **Functionality**:
-  - Enforces authentication before page access
-  - Redirects unauthenticated users to \`/login\`
-  - Maintains session state
-- **Status**: FULLY FUNCTIONAL
-
-#### Portfolio Metrics (Executive Overview Cards)
-- **Hook**: \`usePortfolioMetrics(companyId)\`
-- **File**: \`src/hooks/useChairmanData.ts\`
-- **Database Tables**: \`ventures\`, \`companies\`
-- **Data Fetched**:
-  - Active ventures count (status = 'active')
-  - Final stage ventures (stage IN final_polish, customer_success, venture_active)
-  - At-risk ventures (status IN paused, cancelled)
-  - Portfolio value calculation ($1M average per venture)
-  - Success rate percentage (active/completed vs total)
-- **Refresh**: React Query with 5-minute stale time
-- **Filtering**: Company-specific or "all" aggregation
-- **Status**: CONNECTED (calculated from real venture data)
-
-#### Executive Alerts (Priority Alerts Component)
-- **Hook**: \`useExecutiveAlertsUnified(companyId)\`
-- **Base Hook**: \`useUnifiedNotifications(companyId)\`
-- **File**: \`src/hooks/useUnifiedNotifications.ts\`
-- **Data Sources**:
-  1. **chairman_feedback** table - Strategic feedback requiring attention
-  2. **ventures** table - Paused/cancelled venture alerts
-  3. **compliance_violations** table - Governance violations
-  4. **System notifications** - Hard-coded maintenance messages
-- **Refresh**: 30-second auto-polling
-- **Filtering**: Critical + High urgency only, top 10 items
-- **Features**:
-  - Urgency-based color coding (red/orange/yellow/blue)
-  - Category icons (financial, risk, operational, strategic, opportunity)
-  - Action required flagging
-  - Relative timestamps (formatDistanceToNow)
-- **Status**: FULLY FUNCTIONAL
-
-#### Company Management
-- **Hook**: \`useCompanies()\`
-- **File**: \`src/hooks/useChairmanData.ts\`
-- **Database**: \`companies\` table
-- **Fields**: id, name, description, logo_url, website, industry, settings, created_at
-- **Refresh**: 5-minute stale time
-- **Status**: CONNECTED
-
-#### Company Selector State Management
-- **Component**: \`CompanySelector\`
-- **State**: \`selectedCompany\` (useState in AuthenticatedLayout)
-- **Propagation**: Props passed to child components
-- **Status**: IMPLEMENTED (needs propagation verification in testing)
-
-### PARTIALLY IMPLEMENTED / NEEDS VERIFICATION
-
-#### Performance Drive Cycle
-- **Component**: \`PerformanceDriveCycle\`
-- **File**: \`src/components/chairman/PerformanceDriveCycle.tsx\`
-- **Expected Hook**: Likely uses \`usePortfolioMetrics\` or custom hook
-- **Status**: VERIFY - Component exists, backend connection unknown
-
-#### AI Insights Engine
-- **Component**: \`AIInsightsEngine\`
-- **File**: \`src/components/chairman/AIInsightsEngine.tsx\`
-- **Expected Data**: AI-generated strategic recommendations
-- **Status**: VERIFY - May be mock data or connected to AI service
-
-#### Synergy Opportunities
-- **Component**: \`SynergyOpportunities\`
-- **File**: \`src/components/chairman/SynergyOpportunities.tsx\`
-- **Expected Hook**: Likely analyzes ventures for synergies
-- **Status**: VERIFY - Algorithm implementation unknown
-
-#### Venture Portfolio Overview (Portfolio Tab)
-- **Component**: \`VenturePortfolioOverview\`
-- **Expected Data**: Detailed venture listings with filters
-- **Status**: VERIFY - Component exists, integration level unknown
-
-#### Strategic KPI Monitor (KPIs Tab)
-- **Component**: \`StrategicKPIMonitor\`
-- **Expected Data**: Custom KPIs with targets and trends
-- **Status**: VERIFY - Component exists, data source unknown
-
-#### Financial Analytics (Financial Tab)
-- **Component**: \`FinancialAnalytics\`
-- **Expected Data**: Revenue, costs, ROI, financial forecasts
-- **Status**: VERIFY - Component exists, integration unknown
-
-#### Operational Intelligence (Operations Tab)
-- **Component**: \`OperationalIntelligence\`
-- **Expected Data**: Resource utilization, team metrics, operational KPIs
-- **Status**: VERIFY - Component exists, integration unknown
-
-### STUBBED / NOT IMPLEMENTED
-
-#### Portfolio Performance Summary - Stage Counts
-- **Location**: \`ChairmanDashboard.tsx\` lines 267-287
-- **Issue**: Hard-coded venture stage distribution
-- **Stubbed Values**:
-  \`\`\`typescript
-  <div className="text-2xl font-bold">23</div>  // Ideation
-  <div className="text-2xl font-bold">31</div>  // Validation
-  <div className="text-2xl font-bold">28</div>  // Development
-  <div className="text-2xl font-bold">34</div>  // Launch
-  <div className="text-2xl font-bold">11</div>  // Growth
-  \`\`\`
-- **Required**: Dynamic query from ventures table grouped by stage
-- **Priority**: HIGH - Displays inaccurate data
-
-#### Team Utilization Preview
-- **Location**: \`ChairmanDashboard.tsx\` lines 291-320
-- **Issue**: Hard-coded utilization percentages
-- **Stubbed Values**:
-  - Development: 87%
-  - Strategy: 73%
-- **Required**: Connect to resource allocation or team management tables
-- **Priority**: HIGH - Displays inaccurate data
-
-#### Strategic Decision Support (Intelligence Tab)
-- **Location**: \`ChairmanDashboard.tsx\` lines 351-367
-- **Issue**: Placeholder with "Coming in next update" message
-- **UI Present**: Card with empty state icon (Gauge)
-- **Backend**: NOT IMPLEMENTED
-- **Options**:
-  1. Implement the feature
-  2. Remove the placeholder tab content
-  3. Gray out/disable the tab
-- **Priority**: MEDIUM - Placeholder is honest about status
-
----
-
-## 3. UI/UX ASSESSMENT
-
-### Design System Compliance
-
-#### Component Library (Shadcn/UI)
-- **Card System**: Consistent use of Card, CardHeader, CardTitle, CardContent
-- **Badges**: Proper variant system (default, destructive, secondary, outline)
-- **Buttons**: Size and variant props correctly applied
-- **Tabs**: TabsList, TabsTrigger, TabsContent with proper ARIA
-- **Progress Bars**: Uniform styling with venture color scheme
-- **ScrollArea**: Used for overflow handling in alerts
-
-#### Color Palette
-- **Primary Colors**: venture-blue, venture-success, venture-danger (from tailwind.config.ts)
-- **Alert Urgency Colors**:
-  - Critical: red-500 border, red-50 background, red-100/red-800 badge
-  - High: orange-500 border, orange-50 background, orange-100/orange-800 badge
-  - Medium: yellow-500 border, yellow-50 background, yellow-100/yellow-800 badge
-  - Low: blue-500 border, blue-50 background, blue-100/blue-800 badge
-- **Semantic Usage**: Colors map correctly to urgency levels
-- **Consistency**: venture-* colors used throughout metrics and charts
-
-#### Typography Hierarchy
-- **Page Title**: text-3xl font-bold with gradient (from-primary to-primary/70)
-- **Metric Values**: text-2xl font-bold
-- **Card Titles**: text-sm font-medium (may be too small - see issues below)
-- **Alert Titles**: text-sm font-medium leading-tight
-- **Alert Descriptions**: text-xs text-muted-foreground
-- **Body Text**: Consistent use of muted-foreground for secondary text
-
-#### Spacing & Layout
-- **Grid System**: Responsive grid (cols-1 -> md:cols-2 -> lg:cols-4)
-- **Gap Spacing**: Consistent gap-4 and gap-6 usage
-- **Card Padding**: Standard CardHeader and CardContent padding
-- **Section Spacing**: space-y-6 between major sections
-
-### UI/UX Issues Identified
-
-#### Issue 1: Alert Card Title Size
-- **Location**: ExecutiveAlerts component
-- **Problem**: Alert titles use text-sm which may be too small for scan-ability
-- **Impact**: Reduces readability for executives quickly scanning alerts
-- **Recommendation**: Consider text-base or text-md for titles
-- **Priority**: MEDIUM
-
-#### Issue 2: Priority Alerts Fixed Height
-- **Location**: \`ExecutiveAlerts.tsx\` line 126
-- **Problem**: \`<ScrollArea className="h-80">\` creates fixed 320px height
-- **Impact**:
-  - May cut off alerts with long content
-  - Wastes space when few alerts present
-  - Doesn't adapt to content needs
-- **Related**: This is the issue noted in MANUAL-DASHBOARD-MG5GGDV0
-- **Recommendation**: Use min-h-64 max-h-96 or dynamic height calculation
-- **Priority**: HIGH
-
-#### Issue 3: Export/Configure Button Functionality Unknown
-- **Location**: \`ChairmanDashboard.tsx\` lines 184-191
-- **Problem**: Buttons present but no onClick handlers visible in code
-- **Impact**: May be non-functional placeholders
-- **Testing Required**: Verify buttons do something or show appropriate disabled state
-- **Priority**: MEDIUM
-
-#### Issue 4: Hard-Coded Data Misleading
-- **Location**: Portfolio Performance Summary
-- **Problem**: Displays static counts that don't reflect actual data
-- **Impact**: Executives may make decisions based on incorrect information
-- **User Experience**: Undermines trust in the dashboard
-- **Priority**: CRITICAL
-
-#### Issue 5: Tab Content Density Varies
-- **Observation**: Some tabs (Overview) are dense with multiple widgets, others may be sparse
-- **Impact**: Inconsistent information density across views
-- **Testing Required**: Verify all 6 tabs have appropriate content depth
-- **Priority**: LOW (natural variation is acceptable)
-
-### Accessibility Assessment (WCAG 2.1 AA)
-
-#### Implemented Accessibility Features
-- **Skip Navigation**: \`<SkipNavigation />\` present in AuthenticatedLayout
-- **Navigation Announcer**: \`<NavigationAnnouncer />\` for screen readers
-- **Enhanced Keyboard Nav**: \`useEnhancedKeyboardNavigation()\` hook active
-- **Semantic HTML**: Proper heading hierarchy, main element with role="main"
-- **Focus Management**: SidebarTrigger and interactive elements are proper buttons
-- **Responsive Design**: Mobile-first approach with touch-friendly targets
-
-#### Missing Accessibility Features
-
-##### Screen Reader Support
-- **Missing ARIA Labels**:
-  - Export Report button (line 184): \`<Download className="w-4 h-4 mr-2" />\` - icon needs aria-label
-  - Configure button (line 188): \`<Settings className="w-4 h-4 mr-2" />\` - icon needs aria-label
-  - Alert category icons (ExecutiveAlerts.tsx lines 78-92): Icons need aria-labels
-- **Missing ARIA Live Regions**:
-  - Priority Alerts: Should have \`aria-live="polite"\` for real-time updates
-  - Metric cards: Should announce value changes to screen readers
-- **Missing Role Descriptions**:
-  - Tabs: Should have aria-label describing dashboard sections
-  - Metric cards: Should have aria-describedby for trend information
-
-##### Keyboard Navigation
-- **Testing Required**:
-  - [ ] Tab order through all interactive elements
-  - [ ] Focus indicators visible and high-contrast
-  - [ ] No keyboard traps in modals or overlays
-  - [ ] Escape key closes modals/dropdowns
-  - [ ] Arrow keys work in tab navigation
-  - [ ] Enter/Space activates buttons
-
-##### Color Contrast
-- **Testing Required**:
-  - [ ] Badge text on colored backgrounds (red/orange/yellow/blue)
-  - [ ] Alert text on tinted backgrounds (bg-red-50, bg-orange-50, etc.)
-  - [ ] muted-foreground text contrast
-  - [ ] Link colors in all states (default, hover, visited, focus)
-- **Tool**: Use browser DevTools or WAVE extension for automated checks
-- **Target**: 4.5:1 for normal text, 3:1 for large text
-
-### Responsive Design Analysis
-
-#### Breakpoint Strategy
-- **Mobile**: Default (< 640px) - Single column, collapsed sidebar
-- **Tablet**: md (768px) - 2 columns for metric cards
-- **Laptop**: lg (1024px) - 4 columns for metrics, 2-3 columns for content
-- **Desktop**: xl (1280px+) - Full 4-column layout
-
-#### Responsive Components
-- **Sidebar**: Collapses to overlay on mobile (SidebarTrigger)
-- **Header**: Stacks vertically on small screens
-- **Metric Cards**: Grid adapts from 1 -> 2 -> 4 columns
-- **Tabs**: Grid adapts from 2 -> 6 columns with text hiding on sm breakpoints
-
-#### Responsive Concerns
-- **Tab Labels**: Hidden on small screens (\`<span className="hidden sm:inline">\`)
-  - May confuse users if only icons shown
-  - Needs testing for icon-only comprehension
-- **Priority Alerts**: Fixed height may be problematic on mobile (less vertical space)
-- **Table Data**: If present in sub-components, may need horizontal scroll
-- **Long Text**: Alert descriptions and venture names may wrap awkwardly on narrow screens
-
----
-
-## 4. INTEGRATION CHECK
-
-### Frontend-Backend Mapping Table
-
-| Feature | UI Component | Backend Service | Data Source | Integration Status | Verification Needed |
-|---------|--------------|-----------------|-------------|-------------------|-------------------|
-| **Authentication** | ProtectedRoute | Supabase Auth | auth.users | Connected | [ ] Test unauthorized access |
-| **Navigation** | ModernNavigationSidebar | Static config | navigationItems array | Functional | [ ] Verify active state |
-| **Company Filter** | CompanySelector | useCompanies hook | companies table | Connected | [ ] Test filter propagation |
-| **Portfolio Value** | MetricCard | usePortfolioMetrics | ventures table | Connected | [ ] Verify calculation accuracy |
-| **Active Ventures** | MetricCard | usePortfolioMetrics | ventures table | Connected | [ ] Verify count logic |
-| **Success Rate** | MetricCard | usePortfolioMetrics | ventures table | Connected | [ ] Verify percentage calc |
-| **At Risk Count** | MetricCard | usePortfolioMetrics | ventures table | Connected | [ ] Verify risk criteria |
-| **Priority Alerts** | ExecutiveAlerts | useExecutiveAlertsUnified | chairman_feedback, ventures, compliance_violations | Connected | [ ] Test real-time updates |
-| **Performance Cycle** | PerformanceDriveCycle | Unknown hook | Unknown table | Unknown | [X] **VERIFY BACKEND** |
-| **AI Insights** | AIInsightsEngine | Unknown service | AI/ML endpoint? | Unknown | [X] **VERIFY BACKEND** |
-| **Synergy Ops** | SynergyOpportunities | Unknown hook | ventures analysis? | Unknown | [X] **VERIFY BACKEND** |
-| **Stage Counts** | Hard-coded divs | **NONE** | **STUBBED** | Not Connected | [X] **REPLACE WITH DYNAMIC QUERY** |
-| **Team Utilization** | Hard-coded Progress | **NONE** | **STUBBED** | Not Connected | [X] **CONNECT TO RESOURCE DATA** |
-| **Portfolio Tab** | VenturePortfolioOverview | Unknown hook | ventures table? | Unknown | [X] **VERIFY BACKEND** |
-| **KPIs Tab** | StrategicKPIMonitor | Unknown hook | kpis table? | Unknown | [X] **VERIFY BACKEND** |
-| **Financial Tab** | FinancialAnalytics | Unknown hook | financial_data table? | Unknown | [X] **VERIFY BACKEND** |
-| **Operations Tab** | OperationalIntelligence | Unknown hook | operational_metrics? | Unknown | [X] **VERIFY BACKEND** |
-| **Intelligence Tab** | AIInsightsEngine + Placeholder | Partial | Mixed | Mixed | [X] **COMPLETE OR REMOVE PLACEHOLDER** |
-| **Export Report** | Button | Unknown | Unknown | Unknown | [X] **VERIFY FUNCTIONALITY** |
-| **Configure** | Button | Unknown | Unknown | Unknown | [X] **VERIFY FUNCTIONALITY** |
-
-### Cross-Component Communication
-
-#### Verified Communication Paths
-1. **CompanySelector -> ExecutiveOverviewCards**: \`companyId\` prop -> \`usePortfolioMetrics(companyId)\`
-2. **CompanySelector -> ExecutiveAlerts**: \`companyId\` prop -> \`useExecutiveAlertsUnified(companyId)\`
-3. **CompanySelector -> PerformanceDriveCycle**: \`companyId\` prop passed
-4. **CompanySelector -> SynergyOpportunities**: \`companyId\` prop passed
-5. **Tab State -> Content**: Managed by Shadcn Tabs component (controlled)
-
-#### Needs Verification
-- [ ] Does CompanySelector state persist across tab changes?
-- [ ] Do all 6 tabs respect the company filter?
-- [ ] Do metric cards update immediately when company changes?
-- [ ] Are there loading states during company filter transitions?
-
----
-
-## 5. SUB-AGENT RESPONSIBILITIES
-
-### Design Sub-Agent - UI/UX Enhancements
-
-#### Priority 1: Accessibility Compliance (CRITICAL)
-**Add ARIA Labels to Icon-Only Elements**
-- [ ] Export Report button: \`aria-label="Export chairman dashboard report"\`
-- [ ] Configure button: \`aria-label="Configure dashboard settings"\`
-- [ ] Alert category icons: Add descriptive aria-labels (e.g., "Financial risk", "Operational issue")
-- [ ] Metric trend icons: Add aria-labels (e.g., "Trending up 12%")
-- **Files**: \`ChairmanDashboard.tsx\`, \`ExecutiveAlerts.tsx\`
-- **Impact**: Critical for screen reader users
-
-**Implement ARIA Live Regions**
-- [ ] Priority Alerts section: \`<div aria-live="polite" aria-atomic="true">\`
-- [ ] Metric cards: Announce value changes with aria-live
-- [ ] Company selector: Announce filter changes
-- **Purpose**: Real-time updates accessible to screen readers
-- **Impact**: High - core functionality for assistive tech users
-
-**Screen Reader Context**
-- [ ] Add role descriptions to tab sections
-- [ ] Add aria-describedby to metric cards linking to trend descriptions
-- [ ] Ensure tab focus management works correctly
-- **Testing**: Use NVDA/JAWS for validation
-
-#### Priority 2: Fix Priority Alerts Layout (HIGH)
-**Replace Fixed Height with Adaptive Sizing**
-- **Current**: \`<ScrollArea className="h-80">\` (320px fixed)
-- **Proposed**:
-  \`\`\`tsx
-  <ScrollArea className="min-h-64 max-h-[32rem]">
-  {/* Dynamic height based on content, min 256px, max 512px */}
-  \`\`\`
-- **Alternative**: Calculate height based on alert count
-- **Testing**: Verify with 0, 1, 5, 10+ alerts
-- **File**: \`src/components/chairman/ExecutiveAlerts.tsx\` line 126
-
-**Improve Alert Card Layout**
-- [ ] Increase alert title from text-sm to text-base
-- [ ] Add max-width to long text elements to prevent overflow
-- [ ] Test badge wrapping behavior with multiple badges
-- [ ] Ensure consistent spacing between alert cards
-
-#### Priority 3: Responsive Improvements (MEDIUM)
-- [ ] Test tab icon-only mode on mobile (sm breakpoint)
-- [ ] Verify Priority Alerts on mobile devices (portrait/landscape)
-- [ ] Check metric card stacking on narrow screens
-- [ ] Test header component wrapping/overflow on small tablets
-
-#### Priority 4: Color Contrast Validation (MEDIUM)
-- [ ] Run WAVE or axe DevTools on entire dashboard
-- [ ] Test all badge variants against WCAG AA
-- [ ] Verify alert text on tinted backgrounds (bg-red-50, etc.)
-- [ ] Check muted-foreground contrast ratio
-- **Target**: 4.5:1 minimum for normal text
-
-### Database/EXEC Sub-Agents - Backend Integration
-
-#### Priority 1: Replace Hard-Coded Data (CRITICAL)
-**Portfolio Stage Distribution**
-- [ ] Create query to count ventures grouped by stage
-- [ ] Update \`ChairmanDashboard.tsx\` lines 267-287 with dynamic data
-- [ ] Add error handling and loading states
-- **Query Example**:
-  \`\`\`sql
-  SELECT stage, COUNT(*) as count
-  FROM ventures
-  WHERE company_id = ? OR ? = 'all'
-  GROUP BY stage
-  ORDER BY stage
-  \`\`\`
-- **Hook**: Extend \`usePortfolioMetrics\` or create \`useVentureStageDistribution\`
-
-**Team Utilization Metrics**
-- [ ] Identify/create resource allocation or team management table
-- [ ] Query actual team utilization percentages
-- [ ] Update \`ChairmanDashboard.tsx\` lines 298-311 with dynamic data
-- [ ] Consider removing if no real data source exists
-- **Alternative**: Remove this section entirely if not supported by backend
-
-#### Priority 2: Verify Sub-Component Integrations (HIGH)
-**Components Needing Backend Verification**:
-1. **PerformanceDriveCycle** (\`src/components/chairman/PerformanceDriveCycle.tsx\`)
-   - [ ] Review component source code
-   - [ ] Identify data hooks/queries
-   - [ ] Verify data accuracy
-   - [ ] Add to integration status table
-
-2. **AIInsightsEngine** (\`src/components/chairman/AIInsightsEngine.tsx\`)
-   - [ ] Check if connected to AI/ML service or using mock data
-   - [ ] Verify API endpoints if external service
-   - [ ] Document data refresh strategy
-
-3. **SynergyOpportunities** (\`src/components/chairman/SynergyOpportunities.tsx\`)
-   - [ ] Review synergy detection logic
-   - [ ] Verify ventures analysis queries
-   - [ ] Check calculation accuracy
-
-4. **VenturePortfolioOverview** (Portfolio tab)
-   - [ ] Verify ventures table queries
-   - [ ] Check filtering and sorting functionality
-   - [ ] Validate pagination if present
-
-5. **StrategicKPIMonitor** (KPIs tab)
-   - [ ] Identify KPI data sources
-   - [ ] Verify target vs actual calculations
-   - [ ] Check trend analysis accuracy
-
-6. **FinancialAnalytics** (Financial tab)
-   - [ ] Identify financial data tables
-   - [ ] Verify calculation logic
-   - [ ] Check currency formatting and conversions
-
-7. **OperationalIntelligence** (Operations tab)
-   - [ ] Identify operational metrics sources
-   - [ ] Verify team/resource queries
-   - [ ] Check aggregation logic
-
-#### Priority 3: Implement or Remove Stubs (MEDIUM)
-**Strategic Decision Support (Intelligence Tab)**
-- **Option A**: Implement the feature
-  - Define requirements with stakeholders
-  - Design data model
-  - Implement backend logic
-  - Connect frontend
-- **Option B**: Remove placeholder
-  - Remove card from Intelligence tab
-  - Update UI to show only AIInsightsEngine
-- **Option C**: Disable tab entirely
-  - Gray out Intelligence tab
-  - Show "Coming Soon" indicator
-- **Decision Needed**: Consult with product owner
-
-**Export/Configure Buttons**
-- [ ] Implement export functionality (PDF/Excel report generation)
-- [ ] Implement configure functionality (dashboard customization)
-- [ ] OR: Remove buttons if not planned
-- [ ] OR: Show disabled state with tooltip explaining future feature
-
-#### Priority 4: Performance Optimization (LOW)
-**Alert Polling Frequency**
-- **Current**: 30-second auto-refresh
-- **Consideration**: May be too frequent, causing unnecessary load
-- **Options**:
-  1. Increase to 60 seconds
-  2. Implement WebSocket/Supabase Realtime for instant updates
-  3. Add manual refresh button for user control
-- **Testing**: Monitor database query load
-
-**Database Indexes**
-- [ ] Verify index on ventures.status for alert queries
-- [ ] Verify index on ventures.company_id for filtering
-- [ ] Verify index on chairman_feedback.alert_level
-- [ ] Verify index on compliance_violations.status
-- **Goal**: Sub-100ms query times
-
-### Security Sub-Agent - Access Control & Data Protection
-
-#### Priority 1: Authentication Enforcement (CRITICAL)
-**Protected Route Testing**
-- [ ] Attempt to access \`/chairman\` without authentication
-- [ ] Verify redirect to \`/login\` page
-- [ ] Test session timeout behavior (idle timeout)
-- [ ] Test expired token handling
-- [ ] Verify re-authentication flow
-- **File**: \`ProtectedRoute\` component
-
-**Authorization Checks**
-- [ ] Verify user roles/permissions for chairman console access
-- [ ] Test if non-executive users can access dashboard
-- [ ] Check if certain tabs should be role-restricted
-- **Consideration**: Should all authenticated users see this page?
-
-#### Priority 2: Data Isolation (HIGH)
-**Multi-Company Data Segregation**
-- [ ] Test company filter with multiple companies
-- [ ] Attempt to access data from unauthorized companies
-- [ ] Verify RLS policies on all queried tables
-- [ ] Test "all companies" view authorization
-- **Tables to Verify**: ventures, chairman_feedback, compliance_violations, companies
-
-**Row-Level Security (RLS) Policies**
-- [ ] \`ventures\` table: Verify company_id filtering
-- [ ] \`chairman_feedback\` table: Verify executive-only access
-- [ ] \`compliance_violations\` table: Verify governance role access
-- [ ] \`companies\` table: Verify user_company_access enforcement
-- **Testing**: Use different user accounts with varying permissions
-
-#### Priority 3: Sensitive Data Handling (HIGH)
-**Executive Decision Data**
-- [ ] Verify \`chairman_feedback.executive_decision\` field is protected
-- [ ] Check if sensitive feedback text is sanitized/encrypted
-- [ ] Audit logging for executive decisions
-- **Compliance**: May have regulatory requirements (SOX, etc.)
-
-**Compliance Violation Details**
-- [ ] Verify governance violation data is role-restricted
-- [ ] Check if violation details expose sensitive info
-- [ ] Confirm audit trails for violation access
-- **Consideration**: Some users may only need violation counts, not details
-
-**Portfolio Financial Data**
-- [ ] Verify financial metrics are properly secured
-- [ ] Check if portfolio value calculations expose proprietary data
-- [ ] Confirm investor/stakeholder data isolation
-- **Consideration**: Different detail levels for different roles
-
-#### Priority 4: API Security (MEDIUM)
-**Supabase Client Configuration**
-- [ ] Verify anon key is used (not service role key in frontend)
-- [ ] Check RLS enforcement in all queries
-- [ ] Validate CORS policies
-- [ ] Review Supabase dashboard security settings
-
-**Export Functionality Security** (If Implemented)
-- [ ] Verify exported reports include only authorized data
-- [ ] Check file permissions on generated reports
-- [ ] Implement download link expiration
-- [ ] Add audit logging for report exports
-
----
-
-## TESTING SCOPE FOR MANUAL UAT
-
-### 1. Navigation & Page Load Testing
-
-#### [ ] Sidebar Navigation
-- [ ] Click "Chairman Console" link in sidebar
-- [ ] Verify navigation to \`/chairman\` route
-- [ ] Confirm active state indicator on sidebar item
-- [ ] Test keyboard shortcut (if configured)
-- [ ] Test breadcrumb updates correctly
-- **Expected**: Instant navigation, active state visible, breadcrumb shows "Chairman Console"
-
-#### [ ] Default Redirect
-- [ ] Log in with valid credentials
-- [ ] Observe automatic redirect
-- **Expected**: Redirects to \`/chairman\` (not \`/\` or other page)
-
-#### [ ] Page Load Performance
-- [ ] Measure initial page load time (DevTools Network tab)
-- [ ] Check for loading skeleton states
-- [ ] Verify Suspense fallback displays
-- **Target**: < 3 seconds to interactive on standard connection
-
-#### [ ] Error Handling
-- [ ] Disconnect network and try to load page
-- [ ] Observe error state
-- [ ] Reconnect and verify recovery
-- **Expected**: Graceful error message, retry mechanism
-
-### 2. Authentication & Authorization Testing
-
-#### [ ] Unauthenticated Access
-- [ ] Log out completely
-- [ ] Attempt to navigate directly to \`/chairman\`
-- **Expected**: Redirect to \`/login\` page
-
-#### [ ] Session Timeout
-- [ ] Log in and remain idle for configured timeout period
-- [ ] Attempt to interact with dashboard
-- **Expected**: Session expires, prompt to re-authenticate
-
-#### [ ] Role-Based Access (If Applicable)
-- [ ] Log in with non-executive account (if roles exist)
-- [ ] Attempt to access Chairman Console
-- **Expected**: Access denied or limited view
-
-### 3. Company Filtering Testing
-
-#### [ ] Company Selector Functionality
-- [ ] Open CompanySelector dropdown
-- [ ] Verify all companies are listed
-- [ ] Select a specific company
-- [ ] Observe dashboard update
-- **Expected**: All metrics and alerts filter to selected company
-
-#### [ ] "All Companies" View
-- [ ] Select "All" in CompanySelector
-- [ ] Verify aggregated data displays
-- **Expected**: Metrics sum across all companies
-
-#### [ ] Filter Propagation
-- [ ] Select Company A
-- [ ] Check all 4 metric cards show Company A data
-- [ ] Check Priority Alerts show Company A alerts
-- [ ] Switch to Overview tab widgets
-- [ ] Switch to Portfolio tab
-- [ ] Switch to other tabs (KPIs, Financial, Operations, Intelligence)
-- **Expected**: All components respect selected company filter
-
-#### [ ] Filter Persistence
-- [ ] Select Company B
-- [ ] Switch between tabs
-- [ ] Refresh page
-- **Expected**: Selected company persists across tabs, resets on page refresh
-
-### 4. Executive Overview Cards Testing
-
-#### [ ] Portfolio Value Card
-- [ ] Verify numerical value displays
-- [ ] Check currency formatting (e.g., $42.5M)
-- [ ] Verify trend indicator (up/down arrow)
-- [ ] Check percentage change value
-- [ ] Hover for additional details (if applicable)
-- **Data Validation**: Compare with actual venture count x $1M calculation
-
-#### [ ] Active Ventures Card
-- [ ] Verify count displays
-- [ ] Check sub-value (final stage ventures count)
-- [ ] Verify count matches ventures with status="active"
-- **Data Validation**: Cross-reference with Ventures page count
-
-#### [ ] Success Rate Card
-- [ ] Verify percentage displays
-- [ ] Check trend indicator
-- [ ] Verify percentage change value
-- [ ] Confirm calculation: (active + completed) / total x 100
-- **Data Validation**: Manual calculation from ventures data
-
-#### [ ] At Risk Card
-- [ ] Verify count displays
-- [ ] Check trend indicator (note: increase is negative trend)
-- [ ] Verify change count (not percentage)
-- [ ] Confirm count matches ventures with status IN ('paused', 'cancelled')
-- **Data Validation**: Cross-reference with paused/cancelled ventures
-
-#### [ ] Loading States
-- [ ] Observe skeleton loaders on initial page load
-- [ ] Verify 4 skeleton cards display
-- [ ] Check smooth transition to real data
-- **Expected**: No layout shift when data loads
-
-#### [ ] Empty States
-- [ ] Test with company that has 0 ventures
-- [ ] Verify graceful handling (0 values, not errors)
-- **Expected**: Cards display 0 with appropriate messaging
-
-### 5. Priority Alerts Testing
-
-#### [ ] Alert Display
-- [ ] Verify Priority Alerts card renders
-- [ ] Check alert count badges (Critical, High)
-- [ ] Verify urgency-based color coding
-- [ ] Check category icons (Financial, Risk, Operational, Strategic, Opportunity)
-- [ ] Verify relative timestamps ("2 hours ago")
-- [ ] Check ScrollArea scrollbar appears if >10 alerts
-
-#### [ ] Alert Content
-- [ ] Read alert title (should be clear and scannable)
-- [ ] Read alert description (provides context)
-- [ ] Verify venture/company name badge
-- [ ] Verify urgency level badge
-- [ ] Check "Action Required" button displays when needed
-
-#### [ ] Real-Time Updates
-- [ ] Wait 30 seconds (auto-refresh interval)
-- [ ] Verify new alerts appear (if any)
-- [ ] Check alert count badges update
-- **Testing Tip**: Create a test alert in database to observe update
-
-#### [ ] Interactions
-- [ ] Click "Action Required" button
-- [ ] Verify navigation or action occurs
-- [ ] Click "View All X Alerts" button at bottom
-- [ ] Verify navigation to full notifications page
-- **Expected**: Buttons are functional (not placeholders)
-
-#### [ ] Empty State
-- [ ] Test with company that has no alerts
-- [ ] Verify empty state displays
-- **Expected**: "No alerts at this time" message with icon
-
-#### [ ] Layout Issue Verification (CRITICAL TEST)
-- [ ] Observe Priority Alerts section in 3-column grid
-- [ ] Check if component height fits properly
-- [ ] Test with varying alert counts (1, 5, 10+)
-- [ ] Test with long alert titles and descriptions
-- [ ] Test responsive behavior (tablet, mobile)
-- **Known Issue**: Fixed height may cause layout problems
-- **Expected After Fix**: Dynamic height, no overflow/cutoff
-
-### 6. Tab Navigation Testing
-
-#### [ ] Tab Switching
-- [ ] Click each of the 6 tabs in order
-- [ ] Verify tab content changes
-- [ ] Check active tab indicator (underline/highlight)
-- [ ] Verify no layout shifts during transitions
-- **Tabs**: Overview, Portfolio, KPIs, Financial, Operations, Intelligence
-
-#### [ ] Keyboard Navigation
-- [ ] Tab to tab list (focus on first tab)
-- [ ] Use arrow keys to move between tabs
-- [ ] Press Enter/Space to activate tab
-- [ ] Tab into content area
-- **Expected**: Full keyboard accessibility
-
-#### [ ] Tab Icons
-- [ ] On mobile (< 640px), verify icons display
-- [ ] Check if icon-only labels are comprehensible
-- [ ] On desktop, verify icon + text label display
-- **Consideration**: Icon-only mode may need tooltips
-
-#### [ ] Tab Content Loading
-- [ ] Switch to each tab
-- [ ] Verify content loads (not blank)
-- [ ] Check for loading spinners if async data
-- [ ] Verify error handling if data fails
-- **Expected**: All tabs have meaningful content (no "Coming Soon" except Decision Support)
-
-### 7. Overview Tab Components Testing
-
-#### [ ] Performance Drive Cycle
-- [ ] Verify component renders
-- [ ] Check data visualization accuracy
-- [ ] Test interactions (hover, click, etc.)
-- **Backend Verification Needed**: Confirm data source
-
-#### [ ] AI Strategic Insights
-- [ ] Verify insights display
-- [ ] Check insight quality/relevance
-- [ ] Verify refresh mechanism
-- **Backend Verification Needed**: Real AI vs mock data?
-
-#### [ ] Synergy Opportunities
-- [ ] Verify opportunities display
-- [ ] Check synergy descriptions
-- [ ] Test if clickable (links to ventures)
-- **Backend Verification Needed**: Confirm algorithm
-
-#### [ ] Portfolio Performance Summary
-- [ ] **CRITICAL**: Verify stage counts (Ideation, Validation, Development, Launch, Growth)
-- [ ] **KNOWN ISSUE**: Currently hard-coded (23, 31, 28, 34, 11)
-- [ ] **TEST**: Change venture stages in database, refresh page
-- [ ] **EXPECTED AFTER FIX**: Counts update dynamically
-- [ ] Check gradient bar display
-- [ ] Verify Team Utilization section
-- [ ] **KNOWN ISSUE**: Utilization percentages hard-coded (87%, 73%)
-- [ ] Click "View detailed operations ->" link
-- **Expected After Fixes**: All data is dynamic and accurate
-
-### 8. Portfolio Tab Testing
-
-#### [ ] Content Presence
-- [ ] Switch to Portfolio tab
-- [ ] Verify VenturePortfolioOverview component loads
-- [ ] Check for venture listings
-- [ ] Test filters/sorting (if present)
-- [ ] Test pagination (if present)
-- **Backend Verification Needed**: Confirm integration
-
-#### [ ] Detailed View
-- [ ] Click on individual venture (if clickable)
-- [ ] Verify navigation to venture detail
-- [ ] Test back navigation
-- **Expected**: Deep linking works correctly
-
-### 9. KPIs Tab Testing
-
-#### [ ] Content Presence
-- [ ] Switch to KPIs tab
-- [ ] Verify StrategicKPIMonitor component loads
-- [ ] Check for KPI cards/tables
-- [ ] Verify target vs actual displays
-- [ ] Check trend indicators
-- **Backend Verification Needed**: Confirm data source
-
-#### [ ] KPI Interactions
-- [ ] Test sorting (if present)
-- [ ] Test filtering by KPI type (if present)
-- [ ] Test time period selection (if present)
-- **Expected**: Interactive KPI management
-
-### 10. Financial Tab Testing
-
-#### [ ] Content Presence
-- [ ] Switch to Financial tab
-- [ ] Verify FinancialAnalytics component loads
-- [ ] Check for financial charts/metrics
-- [ ] Verify currency formatting
-- [ ] Check time series data (if present)
-- **Backend Verification Needed**: Confirm integration
-
-#### [ ] Financial Calculations
-- [ ] Verify revenue, costs, ROI calculations
-- [ ] Check for data consistency with other views
-- [ ] Test export/print functionality (if present)
-- **Data Validation**: Cross-check with source financial records
-
-### 11. Operations Tab Testing
-
-#### [ ] Content Presence
-- [ ] Switch to Operations tab
-- [ ] Verify OperationalIntelligence component loads
-- [ ] Check for operational metrics
-- [ ] Verify team/resource utilization displays
-- [ ] Check for operational KPIs
-- **Backend Verification Needed**: Confirm integration
-
-#### [ ] Operational Data
-- [ ] Verify team utilization (should NOT be hard-coded 87%/73%)
-- [ ] Check resource allocation displays
-- [ ] Verify operational efficiency metrics
-- **Expected**: Real operational data, not placeholders
-
-### 12. Intelligence Tab Testing
-
-#### [ ] AI Insights Section
-- [ ] Verify AIInsightsEngine displays
-- [ ] Check insight quality and relevance
-- [ ] Test refresh/regenerate functionality (if present)
-
-#### [ ] Strategic Decision Support (KNOWN PLACEHOLDER)
-- [ ] Verify placeholder card displays
-- [ ] Check "Coming in next update" message
-- [ ] Confirm Gauge icon displays
-- [ ] **DECISION NEEDED**: Implement, remove, or keep placeholder
-- **Expected**: Honest placeholder or functional feature
-
-### 13. Header Components Testing
-
-#### [ ] Global Search
-- [ ] Type query in HeaderSearch input
-- [ ] Verify search results display
-- [ ] Test clicking result to navigate
-- [ ] Test search scope (all entities)
-- **Expected**: Quick navigation to ventures, portfolios, etc.
-
-#### [ ] Breadcrumb Navigation
-- [ ] Verify breadcrumb displays "Chairman Console"
-- [ ] Check if breadcrumb updates when drilling down (if applicable)
-- [ ] Test clicking breadcrumb segments to navigate
-- **Expected**: Accurate location indicator
-
-#### [ ] Keyboard Shortcuts
-- [ ] Click KeyboardShortcuts icon/button
-- [ ] Verify shortcut reference modal opens
-- [ ] Check if shortcuts are documented
-- [ ] Test actual shortcuts (if configured)
-- **Expected**: Help dialog with shortcut list
-
-#### [ ] Dark Mode Toggle
-- [ ] Click DarkModeToggle
-- [ ] Verify theme switches (light <-> dark)
-- [ ] Check all components adapt to theme
-- [ ] Verify persistence (reload page)
-- **Expected**: Smooth theme transition, no broken styles
-
-#### [ ] User Menu
-- [ ] Click UserMenu dropdown
-- [ ] Verify Profile option displays
-- [ ] Verify Settings option displays
-- [ ] Verify Logout option displays
-- [ ] Test logout functionality
-- **Expected**: Dropdown works, logout redirects to login
-
-### 14. Export/Configure Buttons Testing
-
-#### [ ] Export Report Button
-- [ ] Click "Export Report" button
-- [ ] **IF IMPLEMENTED**: Verify report generates
-- [ ] **IF IMPLEMENTED**: Check report format (PDF, Excel, etc.)
-- [ ] **IF IMPLEMENTED**: Verify report contains current filtered data
-- [ ] **IF PLACEHOLDER**: Verify disabled state or "Coming Soon" tooltip
-- **Backend Verification Needed**: Confirm functionality exists
-
-#### [ ] Configure Button
-- [ ] Click "Configure" button
-- [ ] **IF IMPLEMENTED**: Verify configuration modal opens
-- [ ] **IF IMPLEMENTED**: Test dashboard customization options
-- [ ] **IF IMPLEMENTED**: Verify changes save and persist
-- [ ] **IF PLACEHOLDER**: Verify disabled state or "Coming Soon" tooltip
-- **Backend Verification Needed**: Confirm functionality exists
-
-### 15. Responsive Design Testing
-
-#### [ ] Mobile Portrait (375px x 667px)
-- [ ] Verify sidebar collapses to overlay
-- [ ] Check SidebarTrigger (hamburger menu) works
-- [ ] Verify metric cards stack to 1 column
-- [ ] Check tab labels (icon-only on small screens)
-- [ ] Test Priority Alerts scrolling
-- [ ] Verify header components stack/hide appropriately
-- [ ] Check touch targets are >=44px x 44px
-- **Device**: Test on actual mobile device or DevTools emulation
-
-#### [ ] Tablet Portrait (768px x 1024px)
-- [ ] Verify metric cards display in 2 columns
-- [ ] Check sidebar remains visible or easily toggleable
-- [ ] Verify tab labels display (not just icons)
-- [ ] Check Priority Alerts layout
-- [ ] Test all interactions work with touch
-- **Device**: Test on iPad or similar
-
-#### [ ] Laptop (1024px x 768px)
-- [ ] Verify metric cards display in 4 columns
-- [ ] Check sidebar always visible
-- [ ] Verify all tab labels display
-- [ ] Check Priority Alerts in 3-column grid
-- [ ] Test all components fit without horizontal scroll
-- **Device**: Standard laptop screen
-
-#### [ ] Desktop (1920px x 1080px)
-- [ ] Verify layout uses full width appropriately
-- [ ] Check no excessive whitespace
-- [ ] Verify all components scale well
-- [ ] Check font sizes remain readable
-- **Device**: Large monitor
-
-### 16. Accessibility Testing
-
-#### [ ] Screen Reader Testing
-- [ ] Use NVDA (Windows) or VoiceOver (Mac)
-- [ ] Navigate through entire page with Tab key
-- [ ] Verify all interactive elements are announced
-- [ ] Check alert updates are announced (aria-live)
-- [ ] Verify metric card values are announced
-- [ ] Check tab labels are announced correctly
-- **Expected**: Full page navigable and comprehensible via screen reader
-
-#### [ ] Keyboard-Only Navigation
-- [ ] Disconnect mouse/trackpad
-- [ ] Tab through all interactive elements
-- [ ] Verify focus indicators are visible
-- [ ] Test tab switching with arrow keys
-- [ ] Test dropdown navigation with arrow keys
-- [ ] Press Escape to close modals/dropdowns
-- [ ] Test Enter/Space to activate buttons
-- **Expected**: All functionality accessible via keyboard
-
-#### [ ] Color Contrast
-- [ ] Use browser DevTools or WAVE extension
-- [ ] Check all text against backgrounds
-- [ ] Verify badge text on colored backgrounds
-- [ ] Check alert text on tinted backgrounds
-- [ ] Test both light and dark modes
-- **Target**: WCAG AA - 4.5:1 for normal text, 3:1 for large text
-
-#### [ ] Zoom & Text Scaling
-- [ ] Zoom to 200% (Ctrl/Cmd + +)
-- [ ] Verify no horizontal scrolling
-- [ ] Check text remains readable
-- [ ] Verify components don't overlap
-- [ ] Test browser text scaling (Settings)
-- **Expected**: Layout adapts, content remains accessible
-
-### 17. Performance Testing
-
-#### [ ] Load Time
-- [ ] Clear browser cache
-- [ ] Hard reload page (Ctrl+Shift+R)
-- [ ] Measure time to interactive (DevTools Performance tab)
-- **Target**: < 3 seconds on standard connection
-
-#### [ ] Data Refresh Performance
-- [ ] Monitor network requests during 30-second auto-refresh
-- [ ] Check if queries are optimized (no N+1 queries)
-- [ ] Verify no unnecessary re-renders
-- [ ] Use React DevTools Profiler
-- **Expected**: Efficient queries, smooth UI updates
-
-#### [ ] Large Dataset Handling
-- [ ] Test with company that has 100+ ventures
-- [ ] Verify page remains responsive
-- [ ] Check if pagination is implemented
-- [ ] Test scrolling performance
-- **Expected**: No performance degradation with large datasets
-
-#### [ ] Concurrent Users (If Possible)
-- [ ] Simulate multiple users accessing dashboard
-- [ ] Monitor database load
-- [ ] Check for race conditions
-- [ ] Verify data isolation
-- **Expected**: System remains stable under load
-
-### 18. Data Accuracy Testing
-
-#### [ ] Cross-Reference with Source Data
-- [ ] Open Ventures page in another tab
-- [ ] Count actual active ventures
-- [ ] Compare with Active Ventures metric card
-- [ ] **Expected**: Counts match exactly
-
-#### [ ] Company Filter Accuracy
-- [ ] Select Company X
-- [ ] Note Portfolio Value
-- [ ] Manually sum ventures for Company X
-- [ ] **Expected**: Calculated value matches displayed value
-
-#### [ ] Alert Source Verification
-- [ ] Review chairman_feedback table in database
-- [ ] Check for recent critical/high alerts
-- [ ] Verify alerts appear in Priority Alerts section
-- [ ] **Expected**: Database alerts match displayed alerts
-
-#### [ ] Success Rate Calculation
-- [ ] Count ventures with status='active' or 'completed'
-- [ ] Count total ventures
-- [ ] Calculate (successful / total) x 100
-- [ ] **Expected**: Matches Success Rate card value
-
-### 19. Error State Testing
-
-#### [ ] Network Failure
-- [ ] Disconnect network
-- [ ] Observe error messages
-- [ ] Reconnect network
-- [ ] Verify automatic recovery or retry button
-- **Expected**: Graceful degradation, clear error messaging
-
-#### [ ] Invalid Company Selection
-- [ ] Manually set invalid company ID in URL/state
-- [ ] Observe error handling
-- [ ] **Expected**: Fallback to "all companies" or clear error
-
-#### [ ] Empty Data States
-- [ ] Test with new company (no ventures)
-- [ ] Verify empty state messages display
-- [ ] Check for helpful guidance ("Get started" messages)
-- [ ] **Expected**: Friendly empty states, not errors
-
-#### [ ] Query Timeout
-- [ ] Simulate slow network (DevTools throttling)
-- [ ] Observe loading states
-- [ ] Verify timeout handling (if configured)
-- [ ] **Expected**: Loading indicators, timeout errors after reasonable wait
-
-### 20. Security Testing
-
-#### [ ] Unauthorized Access
-- [ ] Log out
-- [ ] Attempt to navigate to \`/chairman\`
-- [ ] **Expected**: Redirect to login, no data visible
-
-#### [ ] Company Data Isolation
-- [ ] Log in as User A (access to Company X only)
-- [ ] Select "All Companies" in filter
-- [ ] **Expected**: See only Company X data, not other companies
-
-#### [ ] SQL Injection Attempts
-- [ ] Enter SQL injection strings in search/filters (e.g., \`'; DROP TABLE ventures;--\`)
-- [ ] **Expected**: Inputs sanitized, no database errors
-
-#### [ ] XSS Attempts
-- [ ] Enter script tags in editable fields (if any)
-- [ ] **Expected**: Scripts sanitized/escaped, not executed
-
----
-
-## PRIORITY ACTIONS SUMMARY
-
-### CRITICAL (Block UAT Pass)
-1. Replace hard-coded portfolio stage counts with dynamic queries
-2. Replace hard-coded team utilization with dynamic data OR remove section
-3. Fix Priority Alerts fixed height layout issue (min-h/max-h solution)
-4. Add missing ARIA labels to all icon-only buttons
-5. Verify authentication enforcement (unauthenticated access redirects)
-6. Verify company data isolation (RLS policies working)
-
-### HIGH (Strongly Recommended)
-1. Implement ARIA live regions for real-time updates
-2. Verify all 6 tab components have backend integration
-3. Test Export/Configure button functionality (implement or remove)
-4. Conduct full keyboard navigation testing
-5. Verify color contrast ratios (WCAG AA compliance)
-6. Test responsive design at all breakpoints
-
-### MEDIUM (Nice to Have)
-1. Decide on Strategic Decision Support: implement, remove, or keep placeholder
-2. Increase alert title font size from text-sm to text-base
-3. Add tooltips to icon-only tab labels on mobile
-4. Optimize alert polling frequency (consider WebSocket)
-5. Add database indexes for frequently queried alert data
-6. Implement audit logging for executive decisions
-
-### LOW (Future Enhancement)
-1. Add dashboard customization features (Configure button)
-2. Implement PDF/Excel export functionality
-3. Add keyboard shortcuts for tab navigation
-4. Enhance empty states with "Get Started" guidance
-5. Add data export functionality for individual sections
-
----
-
-## NOTES FOR QA TEAM
-
-### Testing Environment Setup
-- **Database**: Requires populated ventures, chairman_feedback, compliance_violations tables
-- **Test Users**: Need accounts with varying permissions (executive, non-executive)
-- **Test Companies**: Need multiple companies with varying venture counts
-- **Test Data**: Create test alerts in different urgency levels
-
-### Known Issues to Document
-1. Priority Alerts fixed height may cause layout problems (reported in MANUAL-DASHBOARD-MG5GGDV0)
-2. Portfolio stage counts are hard-coded (do not reflect real data)
-3. Team utilization percentages are hard-coded (do not reflect real data)
-4. Strategic Decision Support is a placeholder (not implemented)
-5. Export/Configure buttons may be non-functional placeholders
-
-### Testing Tools Recommended
-- **Accessibility**: WAVE, axe DevTools, Lighthouse
-- **Screen Readers**: NVDA (Windows), VoiceOver (Mac), JAWS
-- **Performance**: Chrome DevTools Performance tab, React DevTools Profiler
-- **Responsive**: Browser DevTools device emulation + real devices
-- **Network**: Chrome DevTools Network tab, throttling options
-
-### Success Criteria
-- All critical issues resolved
-- No broken functionality (buttons do something or are properly disabled)
-- Data accuracy verified (no hard-coded values displayed as real data)
-- Accessibility baseline met (WCAG 2.1 AA compliance)
-- Responsive design works on mobile, tablet, desktop
-- Performance targets met (< 3 second load time)
-- Security requirements met (authentication, authorization, data isolation)
-
----
-
-**Assessment Completed**: 2025-10-01
-**Next Steps**: Execute manual UAT testing using checklists above
-**Follow-Up**: Document findings, create bug tickets for failures, verify fixes
-**Related Test**: MANUAL-DASHBOARD-MG5GGDV0 (focuses on specific Priority Alerts issue)`;
+// Re-export assembled template for backward compatibility
+export { comprehensiveAssessment } from './sections/index.js';
+
+// Re-export individual sections for granular access
+export {
+  headerSection,
+  pageIntentSection,
+  backendEvaluationSection,
+  uiUxAssessmentSection,
+  integrationCheckSection,
+  subAgentResponsibilitiesSection,
+  testingScopeSection,
+  summarySection,
+} from './sections/index.js';
