+++ scripts/modules/uat-assessment/sections/header.js
@@ -1,0 +1,16 @@
+/**
+ * UAT Assessment Template - Header Section
+ * Contains title, metadata, and assessment type information
+ *
+ * @module uat-assessment/sections/header
+ */
+
+export const headerSection = `# Chairman Console - Comprehensive UI/UX Quality Assessment (Systematic UAT)
+
+**Test ID**: TEST-NAV-001
+**Section**: chairman-console
+**Route**: \`/chairman\`
+**Assessment Date**: 2025-10-01
+**Assessment Type**: Full Page Systematic UAT Evaluation
+
+---`;
